a hack to work around what appears to be just a wrong way of doing things, but will ask Noel later.
The basic problem is that (with-check-info* '() (lambda () ...whatever...)) can, in some cases,
double the information that is recorded in a failed test case. Probably the right thing is
to use a parameter or something instead of using continuation marks directly.

original commit: 8ece97219a4e204af9459e27b676e62bd67fd52d