Add an internal location-info struct
This reproduces the behavior that `run-tests` and the default check
handler provide for the `location` info. It changes `check-location` to
automatically wrap the location value in the struct so `define-check`
doesn’t need to change. This breaks clients that construct a check info
with `check-location` and immediately extract the info value field
expecting it to be a `location/c`, but I don’t think anyone does that.

Future work could expose this to rack unit users, but it should
probably be done in terms of the `srcloc` structure instead.