Unify `run-tests` and default check handling of `location` field
This adds a private `location-info` struct alongside `string-info` from #27 to encapsulate rendering a source location with a relative directory path. This obsoletes the logic in both `rackunit/private/format` and `rackunit/text-ui` for handling a check info with key `'location` specially. This is not technically backwards compatible as it changes the observable behavior of `make-check-location`, but it does so in a way that's highly unlikely to affect existing users. See below for details.

* Remove unnecessary check-name? clause of run-tests

The `else` branch at the end of `check-name?` handles this already.

* Change check-info to use struct instead of define-struct

The latter is deprecated. Existing behavior is preserved with
#:constructor-name.

* Allow define-check-type to wrap info values

Needed for a future commit that adds a location-info struct.

* Add an internal location-info struct

This reproduces the behavior that `run-tests` and the default check
handler provide for the `location` info. It changes `check-location` to
automatically wrap the location value in the struct so `define-check`
doesn’t need to change. This breaks clients that construct a check info
with `check-location` and immediately extract the info value field
expecting it to be a `location/c`, but I don’t think anyone does that.

Future work could expose this to rack unit users, but it should
probably be done in terms of the `srcloc` structure instead.

* Bump version number

* Remove text-ui-util-test

This is now an internal implementation detail of
rackunit/private/check-info. Future work might test `location-info` and
`info-value->string` more thoroughly.

* Fix test-test w.r.t. location-info

Also trims some whitespace (auto formatted by Atom)

* Fix bug in prop:exn:srclocs

* Re-add tests for trim-current-directory in check-info tests

* Simplify default info-value printing logic