Implement check-info with a parameter (#41)
This replaces the previous implementation that directly manipulated
continuation marks. After this change, errors raised in tests printed
with run-tests no longer print check info from where the error was
raised. Tests run outside of run-tests never did that in the first
place and nobody complained, so I don’t think it’s a big loss compared
to the simpler implementation. A more principled way to do that would
be for define-check to wrap unexpected exceptions in an exception that
captures check info.