Add a verbose-info type
This lets RackUnit mark some info types as “verbose only”, and they’ll
only show up when run-tests is used in ‘verbose mode. Normal checks
evaluated in DrRacket won’t show verbose info.

The only info type marked as verbose is expression which is typically
redundant in DrRacket, since the check failure message is supposed to
have an exception srcloc pointing to the failure expression. Checks run
in CI or from the command line won’t show the expression, but the error
message does already include the file and line number.

This does not expose the verbose-info struct to RackUnit users.
Instead, it is used internally to remove the now-redundant verbose-info
logic in rackunit/text-ui. With this change, rackunit/text-ui and
normal check evaluation are at “feature parity” in terms of how they
treat check failures, and the run-tests implementation can be stripped
down and the monad logic removed (see #35).