Replace rackunit/text-ui with a simpler implementation
Closes #35, depends on #45

This throws the old implementation out, which was implemented with a
strange ad-hoc dynamic (and impure!) monad-like abstraction backed with
a hash map and lots of lambdas. Now a simple struct is used to manage
the test fold state. Additionally, all duplicated logic in
rackunit/private/format is removed; rackunit/text-ui now contains no
knowledge of the default output format of check messages.