Remove check info stack sorting
The new parameter representation of infos ensures that the built-in
`define-check` infos appear in order before any user-added infos. The
one exception is that the `expression` info now appears before the
`actual` or `expected` infos, but since it’s only present in verbose
mode I don’t think that’s much of an issue.