Provide access to the impl name via syntax-time values.
`define-check` now produces a compile-time value which is a
`check-transformer`. This also has an API to get the name of
the impl function which accepts source location info.

This is needed to fix `typed/rackunit` source locations,
because `typed/rackunit` needs to re-implement the check
macros that wrap these impl functions.