Update test to reflect changes in #123 and #109
Sine we "treat exceptions thrown by arguments to check as test
failures", and test failures contain "params" clause, it makes sense
for the reported message to contain the clause as well.
The clause is also useful in itself.

Use procedure-rename to avoid names
like #<procedure:.../path/to/file.rkt:x:y> which is brittle and
especially too sensitive to changes.