Fix incorrect usage of optional message in rackunit-test
The tests either meant to use `check =` without an epsilon, or
`check-=` with an epsilon. Currently, it uses `check =` with an epsilon,
so the epsilon becomes a message. This PR fixes the tests.

Furthermore, there should be a guard to check that the optional message
argument is indeed either `#f` or `string?`. This PR also does that.